titanic/requirements.txt
fastapi
scikit-learn
unicorn
pydantic
mlflow==2.1.1
python-dotenv

